assorted fixes:  * rename c/cp to cpu/proc  * rename cpu.context to cpu.scheduler  * fix some comments  * formatting for printout

+++ mmu.h
@@ -44,8 +44,8 @@ struct segdesc {
 // Normal segment
 #define SEG(type, base, lim, dpl)                                              \
   (struct segdesc){((lim) >> 12) & 0xffff,                                     \
-                   (base) & 0xffff,                                            \
-                   ((base) >> 16) & 0xff,                                      \
+                   (uint)(base) & 0xffff,                                      \
+                   ((uint)(base) >> 16) & 0xff,                                \
                    type,                                                       \
                    1,                                                          \
                    dpl,                                                        \
@@ -59,8 +59,8 @@ struct segdesc {
 
 #define SEG16(type, base, lim, dpl)                                            \
   (struct segdesc){(lim) & 0xffff,                                             \
-                   (base) & 0xffff,                                            \
-                   ((base) >> 16) & 0xff,                                      \
+                   (uint)(base) & 0xffff,                                      \
+                   ((uint)(base) >> 16) & 0xff,                                \
                    type,                                                       \
                    1,                                                          \
                    dpl,                                                        \
